[backend] move files to app folder
