+++ bankme-backend/src/app/app.module.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
-import { DbModule } from './db/db.module';
+import { DbModule } from '../db/db.module';
 import { PayableModule } from './payable/payable.module';
 import { AssignorModule } from './assignor/assignor.module';
 import { UserModule } from './user/user.module';
 import { CryptoModule } from './crypto/crypto.module';
 import { AuthModule } from './auth/auth.module';
-import { validate } from './config/env.validation';
+import { validate } from '../config/env.validation';
 
 @Module({
   imports: [
