+++ bankme-backend/src/app/auth/auth.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { JwtModule } from '@nestjs/jwt';
-import { UserModule } from 'src/user/user.module';
-import { CryptoModule } from 'src/crypto/crypto.module';
+import { UserModule } from 'src/app/user/user.module';
+import { CryptoModule } from 'src/app/crypto/crypto.module';
 import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
 import { AuthGuard } from './guards/auth.guard';
