+++ bankme-backend/src/app/auth/dto/auth.dto.ts
