+++ bankme-backend/src/app/auth/guards/auth.guard.ts
@@ -7,7 +7,7 @@ import {
 import { Reflector } from '@nestjs/core';
 import { JwtService } from '@nestjs/jwt';
 import { Request } from 'express';
-import { IS_PUBLIC_KEY } from '../../utils/isPublic.decorator';
+import { IS_PUBLIC_KEY } from '../../../utils/isPublic.decorator';
 
 @Injectable()
 export class AuthGuard implements CanActivate {
