+++ bankme-backend/src/app/payable/payable.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { DbModule } from 'src/db/db.module';
-import { AssignorModule } from 'src/assignor/assignor.module';
+import { AssignorModule } from 'src/app/assignor/assignor.module';
 import { PayableController } from './payable.controller';
 import { PayableService } from './payable.service';
 import { PayableRepository } from './payable.repository';
