+++ bankme-backend/src/app/payable/payable.service.spec.ts
@@ -2,7 +2,7 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { NotFoundException } from '@nestjs/common';
 import { CreatePayableDto, UpdatePayableDto } from './dto/payable.dto';
 import { PayableRepository } from './payable.repository';
-import { AssignorService } from 'src/assignor/assignor.service';
+import { AssignorService } from 'src/app/assignor/assignor.service';
 import { PayableService } from './payable.service';
 
 describe('PayableService', () => {
