+++ bankme-backend/src/app/user/user.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { DbModule } from 'src/db/db.module';
-import { CryptoModule } from 'src/crypto/crypto.module';
+import { CryptoModule } from 'src/app/crypto/crypto.module';
 import { UserController } from './user.controller';
 import { UserService } from './user.service';
 import { UserRepository } from './user.repository';
